mattg: add script for switching B+G pixels to value of nearest red neighbor, and separate canvas code into different script
